requirements.txt
torch
pytorch-lightning==2.4.0
hydra-core==1.3.2
omegaconf
pandas
scikit-learn
seaborn
matplotlib
joblib
torchmetrics
esm
transformers
pyarrow
openpyxl
